pyproject.toml
[tool.poetry]
name = "atom"
version = "0.1.0"
description = ""
authors = ["MagdaSlifierz <m.slifierz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.24"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.5.3"}
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
